Simplify supervision structure
Nested Game.Supervisor makes no sense
after getting rid of tasks.
